UndoRedoManager now has setter and getter method for the data that is going to be undone
